BACKEND/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pandas
numpy
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-groq